# libct: speedup process.Env handling