@kolyshkin 2024-07-01
> Nothing in runtime spec says that the environment needs to be deduplicated, or the order of preference (whether the first or the last value of a variable with the same name is to be used). In C (Linux/glibc), the first value is used. In Go, it's the last one.

I was not entirely correct here. Let me rephrase this:

1. When the environment is set using glibc's `setenv(3)` or `putenv(3)`, or Go's `os.Putenv`, the value from the latter call for a given key takes precedence. This is obvious.
2. If the existing environment is not deduplicated (such as when you supply it directly to execve), both glibc's `getenv(3)` and Go's `os.Getenv` return the value of the first element.
3. When using a higher level primitives (from `os/exec`) in Go, the environment passed is deduplicated in a way that the last value of a key is used. In glibc, there is no such deduplication (as far as I can see), and thus the first value of the key will be used. This is the difference between Go and C.

----

On an unrelated note, I also took a look at crun and it seems it is following runc logic, calling `clearenv(3)` followed by `putenv(3)` for every item in `Process.Env`, when using `environ` for execve -- which result in the latter value of the same env var being used. There is probably no room for a similar optimization as `putenv` in glibc is not thread safe and there are no mutexes to be held etc.