@lifubang 2024-09-03
> PTAL @opencontainers/runc-maintainers

LGTM, it’s a big step.
But as I mentioned in https://github.com/opencontainers/runc/pull/4325#discussion_r1656699472, I suggest to add these envs to hook’s process execution before we have a conclusion for #4323.