@kolyshkin 2024-10-21
OTOH I think this is ready for more reviews and potential inclusion.