@cyphar 2024-12-04
FWIW, I sent a patch to improve the performance of `os.Clearenv` (with the patch it's ~30% faster for an environment with 1000 variables set), though this is fixing other issues in addition to the `os.Clearenv` one. https://github.com/golang/go/pull/70672