@kolyshkin 2024-12-05
I've updated the patch (with somewhat better commit message and `prepareEnv` documentation. I think we can go ahead and merge this for v1.3, and leave the question raised in https://github.com/opencontainers/runc/pull/4325#issuecomment-2327591419 for later.