@lifubang 2024-12-05
> and leave the question raised in [#4325 (comment)](https://github.com/opencontainers/runc/pull/4325#issuecomment-2327591419) for later

But it's a break change for `StartContainer` hook, though this hook may be not wide spread used in practice.
Maybe we should add this break to `CHAGELOG` because we may forget it when making a release.